# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OntoConsensus API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input or username taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Logged out
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/api-key:
    put:
      operationId: updateApiKey
      tags: [auth]
      summary: Save or replace the OpenAI API key on the current account
      description: >-
        Every project this user creates uses this key for its AI moderator.
        Never returned by any endpoint once saved -- only whether one is
        configured.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApiKeyInput"
      responses:
        "200":
          description: API key saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /projects/{id}/export-conversation:
    get:
      operationId: exportConversation
      tags: [moderator]
      summary: >-
        Export the full moderator conversation history (transcripts, system
        events, and interventions) as flat, analysis-ready JSON. Unlike
        /export, available at any point in the project's lifecycle -- not
        gated on full agreement.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Conversation export file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationExportPayload"
  /projects/{id}/moderator:
    get:
      operationId: getModeratorStatus
      tags: [moderator]
      summary: >-
        Get whether the AI moderator is on for the CURRENT member, and
        whether it's usable at all for this project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Moderator status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModeratorStatus"
    put:
      operationId: configureModerator
      tags: [moderator]
      summary: >-
        Turn the AI moderator on for the CURRENT member only (their own mic
        and transcript), using the project creator's saved OpenAI API key.
        Does not affect any other member.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModeratorConfigInput"
      responses:
        "200":
          description: Moderator enabled for this member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModeratorStatus"
  /projects/{id}/moderator/disable:
    post:
      operationId: disableModerator
      tags: [moderator]
      summary: Turn the AI moderator off for the CURRENT member only
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Moderator disabled for this member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModeratorStatus"
  /projects/{id}/moderator/messages:
    get:
      operationId: listModeratorChatMessages
      tags: [moderator]
      summary: >-
        Full persisted moderator chat history for this project (intro,
        system, transcript, and intervention messages), visible to every
        member regardless of their own mic state.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Moderator chat history
          content:
            application/json:
              schema:
                type: object
                required: [messages]
                properties:
                  messages:
                    type: array
                    items:
                      $ref: "#/components/schemas/ModeratorChatMessage"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    AuthCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
        password:
          type: string
          minLength: 6
          maxLength: 128
    RegisterInput:
      type: object
      required: [username, password, apiKey]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
        password:
          type: string
          minLength: 6
          maxLength: 128
        apiKey:
          type: string
          minLength: 1
          description: >-
            OpenAI API key for this account. Required at registration;
            stored encrypted, never returned by any endpoint once saved.
            Every project this user creates uses this key for its AI
            moderator. Can be changed later from the dashboard.
    UpdateApiKeyInput:
      type: object
      required: [apiKey]
      properties:
        apiKey:
          type: string
          minLength: 1
    User:
      type: object
      required: [id, username, apiKeyConfigured]
      properties:
        id:
          type: integer
        username:
          type: string
        apiKeyConfigured:
          type: boolean
          description: True once this account has an OpenAI API key saved.
    AuthResponse:
      type: object
      required: [id, username, token, apiKeyConfigured]
      properties:
        id:
          type: integer
        username:
          type: string
        token:
          type: string
          description: >-
            Bearer token to send as `Authorization: Bearer <token>` on
            subsequent requests. Stored client-side per browser tab (not a
            shared cookie), so multiple accounts can be signed in
            simultaneously in different tabs of the same browser.
        apiKeyConfigured:
          type: boolean
          description: True once this account has an OpenAI API key saved.
...
    ModeratorStatus:
      type: object
      required: [active, configured]
      properties:
        active:
          type: boolean
          description: >-
            Whether the AI moderator is currently on for the CURRENT member
            (their own mic + transcript). Independent of every other member.
        configured:
          type: boolean
          description: >-
            True once the project creator's account has an OpenAI API key
            saved -- this is what actually lets the moderator run.
    ModeratorConfigInput:
      type: object
      properties:
        model:
          type: string
          description: >-
            Exact OpenAI model id to use for AI-generated summaries. Defaults
            to the project's already-configured model, or "gpt-5.6-luna" the
            first time a project is configured.
    ModeratorInterventionEntry:
      type: object
      required: [text, by]
      properties:
        text:
          type: string
        by:
          type: array
          items:
            type: string
          description: Usernames of every member who made this same point, in the order first raised.
    ModeratorChatMessage:
      type: object
      required: [id, type, userId, username, colorSlot, content, matched, className, propertyName, classId, propertyId, examples, counterexamples, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [intro, system, transcript, intervention, export_notice]
          description: >-
            "intro" is the moderator's one-time welcome message. "system" is
            an ephemeral-looking but persisted event like a member enabling
            their mic. "transcript" is a live-transcribed speech chunk from a
            member. "intervention" is a stalled-discussion analysis posted by
            the moderator itself. "export_notice" is the one-time notice
            posted the first time a project is exported, stating that
            consensus was reached and the AI moderator has stopped for good;
            it is always recorded (so it appears in the conversation-export
            file) but only rendered in the live chat UI for AI-enabled
            projects.
        userId:
          type: integer
          nullable: true
          description: Null for moderator-authored messages (intro, intervention).
        username:
          type: string
          nullable: true
        colorSlot:
          type: integer
          nullable: true
        content:
          type: string
        matched:
          type: boolean
          nullable: true
          description: >-
            Only meaningful for "intervention" messages -- whether the
            discussion resolved to a real class/property in this project's
            workspace.
        className:
          type: string
          nullable: true
        propertyName:
          type: string
          nullable: true
        classId:
          type: integer
          nullable: true
        propertyId:
          type: integer
          nullable: true
        examples:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/ModeratorInterventionEntry"
          description: >-
            Structured "for keeping it" points, only for a matched
            intervention message -- what the UI renders as the green
            supporter-pill card.
        counterexamples:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/ModeratorInterventionEntry"
          description: >-
            Structured "for removing it" points, only for a matched
            intervention message -- what the UI renders as the red
            supporter-pill card.
        createdAt:
          type: string
          format: date-time
...
    ConversationExportMember:
      type: object
      required: [userId, username, colorSlot]
      properties:
        userId:
          type: integer
        username:
          type: string
        colorSlot:
          type: integer
    ConversationExportMeta:
      type: object
      required: [projectId, projectName, exportedAt, memberCount, members]
      properties:
        projectId:
          type: integer
        projectName:
          type: string
        exportedAt:
          type: string
          format: date-time
        memberCount:
          type: integer
          description: The project's specified number of members (not just however many have joined).
        members:
          type: array
          items:
            $ref: "#/components/schemas/ConversationExportMember"
    ConversationExportIntervention:
      type: object
      required: [matched, classId, propertyId, className, propertyName, examples, counterexamples]
      properties:
        matched:
          type: boolean
          description: Whether the discussion resolved to a real class/property in this project's workspace.
        classId:
          type: integer
          nullable: true
        propertyId:
          type: integer
          nullable: true
        className:
          type: string
          nullable: true
        propertyName:
          type: string
          nullable: true
        examples:
          type: array
          items:
            $ref: "#/components/schemas/ModeratorInterventionEntry"
        counterexamples:
          type: array
          items:
            $ref: "#/components/schemas/ModeratorInterventionEntry"
    ConversationExportMessage:
      type: object
      required: [sequence, id, type, timestamp, speakerUserId, speakerUsername, content, intervention]
      properties:
        sequence:
          type: integer
          description: 1-based position in the conversation, in chronological order.
        id:
          type: integer
        type:
          type: string
          enum: [intro, system, transcript, intervention, export_notice]
        timestamp:
          type: string
          format: date-time
        speakerUserId:
          type: integer
          nullable: true
          description: Null for moderator-authored messages (intro, intervention).
        speakerUsername:
          type: string
          nullable: true
        content:
          type: string
          description: Plain-text content -- always populated, even for a structured intervention.
        intervention:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/ConversationExportIntervention"
          description: Present only for type "intervention"; null for every other message type.
    ConversationExportPayload:
      type: object
      required: [meta, messageCount, messages]
      properties:
        meta:
          $ref: "#/components/schemas/ConversationExportMeta"
        messageCount:
          type: integer
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ConversationExportMessage"