# lib/api-zod/src/generated/types/authCredentials.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OntoConsensus API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthCredentials {
  /**
     * @minLength 3
     * @maxLength 32
     */
  username: string;
  /**
     * @minLength 6
     * @maxLength 128
     */
  password: string;
}