# lib/api-zod/src/generated/types/conversationExportIntervention.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OntoConsensus API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ModeratorInterventionEntry } from './moderatorInterventionEntry';

export interface ConversationExportIntervention {
  /** Whether the discussion resolved to a real class/property in this project's workspace. */
  matched: boolean;
  classId: number | null;
  propertyId: number | null;
  className: string | null;
  propertyName: string | null;
  examples: ModeratorInterventionEntry[];
  counterexamples: ModeratorInterventionEntry[];
}