# lib/api-zod/src/generated/types/moderatorConfigInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OntoConsensus API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ModeratorConfigInput {
  /** Exact OpenAI model id to use for AI-generated summaries. Defaults to the project's already-configured model, or "gpt-5.6-luna" the first time a project is configured. */
  model?: string;
}